Add files via upload
This project is based on a math puzzle that uses the stack data structure. The objective is to move the leftmost disks to the rightmost stacks. They have to be in the order that they started as or else you will be prompted with a try again message. Have fun playing.